requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Admin==1.6.1
SQLAlchemy-JSON==0.5.0
Werkzeug==2.3.7
email-validator==2.0.0.post2
WTForms==3.0.1
